separate report filters for future use
